@@ -20,29 +20,30 @@
   "private": true,
   "license": "See in LICENSE",
   "devDependencies": {
-    "@angular/animations": "12.1.2",
-    "@angular-builders/custom-webpack": "12.1.0",
-    "@angular/cdk": "12.1.2",
-    "@angular/cli": "12.1.2",
-    "@angular/common": "12.1.2",
-    "@angular/compiler": "12.1.2",
-    "@angular/compiler-cli": "12.1.2",
-    "@angular/core": "12.1.2",
-    "@angular/localize": "12.1.2",
-    "@angular-devkit/build-angular": "12.1.2",
-    "@angular-devkit/build-optimizer": "0.1201.2",
-    "@angular-devkit/core": "12.1.2",
-    "@angular/forms": "12.1.2",
-    "@angular/language-service": "12.1.2",
-    "@angular/platform-browser": "12.1.2",
-    "@angular/platform-browser-dynamic": "12.1.2",
-    "@angular/router": "12.1.2",
+    "@angular/animations": "12.2.3",
+    "@angular-builders/custom-webpack": "12.1.1",
+    "@angular/cdk": "12.2.3",
+    "@angular/cli": "12.2.3",
+    "@angular/common": "12.2.3",
+    "@angular/compiler": "12.2.3",
+    "@angular/compiler-cli": "12.2.3",
+    "@angular/core": "12.2.3",
+    "@angular/localize": "12.2.3",
+    "@angular-devkit/build-angular": "12.2.3",
+    "@angular-devkit/build-optimizer": "0.1202.3",
+    "@angular-devkit/core": "12.2.3",
+    "@angular/forms": "12.2.3",
+    "@angular/language-service": "12.2.3",
+    "@angular/platform-browser": "12.2.3",
+    "@angular/platform-browser-dynamic": "12.2.3",
+    "@angular/router": "12.2.3",
     "@ng-bootstrap/ng-bootstrap": "10.0.0",
-    "@ngtools/webpack": "12.1.2",
+    "@ngtools/webpack": "12.2.3",
     "@ngrx/effects": "10.0.0",
     "@ngrx/router-store": "10.0.0",
     "@ngrx/store": "10.0.0",
     "@ngrx/store-devtools": "10.0.0",
+    "@ng-select/ng-select": "^7.2.0",
     "ngx-contextmenu": "6.0.0",
     "angular-confirmation-popover": "6.0.0",
     "angular-notifier": "9.1.0",
