@@ -9,6 +9,7 @@ import { NgbDropdownModule, NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap'
 
 import { DragulaModule } from 'ng2-dragula';
 import { NgxSelectModule } from 'ngx-select-ex';
+import { NgSelectModule } from '@ng-select/ng-select';
 import { NouisliderModule } from 'ng2-nouislider';
 import { ClipboardModule } from 'ngx-clipboard';
 import { ContextMenuModule } from 'ngx-contextmenu';
@@ -24,7 +25,13 @@ import {
     RestoreConfigurationComponent,
     RecoveryModeComponent
 } from './components/device';
-import { KeymapAddComponent, KeymapEditComponent, KeymapHeaderComponent } from './components/keymap';
+import {
+    KeymapAddComponent,
+    KeymapAddEmptyComponent,
+    KeymapEditComponent,
+    KeymapHeaderComponent,
+    KeymapAddSecondaryMenuComponent
+} from './components/keymap';
 import { LayersComponent } from './components/layers';
 import {
     MacroActionEditorComponent,
@@ -52,7 +59,7 @@ import {
 import { CaptureKeystrokeButtonComponent } from './components/popover/widgets/capture-keystroke';
 import { IconComponent } from './components/popover/widgets/icon';
 import { AboutComponent, SettingsComponent, ContributorBadgeComponent } from './components/agent';
-import { SideMenuComponent } from './components/side-menu';
+import { SecondSideMenuContainerComponent, SideMenuComponent } from './components/side-menu';
 import { SvgKeyboardComponent } from './components/svg/keyboard';
 import {
     SvgIconTextKeyComponent,
@@ -158,6 +165,8 @@ import { FirmwareFileUploadComponent } from './components/device/firmware-file-u
         LayersComponent,
         PopoverComponent,
         KeymapAddComponent,
+        KeymapAddEmptyComponent,
+        KeymapAddSecondaryMenuComponent,
         SideMenuComponent,
         KeypressTabComponent,
         KeymapTabComponent,
@@ -209,6 +218,7 @@ import { FirmwareFileUploadComponent } from './components/device/firmware-file-u
         AutoGrowInputComponent,
         HelpPageComponent,
         ExternalUrlDirective,
+        SecondSideMenuContainerComponent,
         SvgSecondaryRoleComponent,
         UdevRulesComponent,
         HighlightArrowComponent,
@@ -228,6 +238,7 @@ import { FirmwareFileUploadComponent } from './components/device/firmware-file-u
             optionValueField: 'id',
             optionTextField: 'text'
         }),
+        NgSelectModule,
         NouisliderModule,
         NotifierModule.withConfig(angularNotifierConfig),
         ConfirmationPopoverModule.forRoot({
