@@ -8,7 +8,6 @@ import {
     createMd5Hash,
     getMd5HashFromFilename,
     HardwareModules,
-    Keymap,
     LEFT_HALF_MODULE,
     PlayMacroAction,
     UHK_60_DEVICE,
@@ -16,8 +15,8 @@ import {
     UserConfiguration
 } from 'uhk-common';
 
+import * as fromDefaultUserConfig from './reducers/default-user-configuration.reducer';
 import * as fromUserConfig from './reducers/user-configuration';
-import * as fromPreset from './reducers/preset';
 import * as fromAppUpdate from './reducers/app-update.reducer';
 import * as fromContributors from './reducers/contributors.reducer';
 import * as autoUpdateSettings from './reducers/auto-update-settings';
@@ -28,7 +27,7 @@ import * as fromUserConfigHistory from './reducers/user-configuration-history.re
 import * as fromSelectors from './reducers/selectors';
 import { initProgressButtonState } from './reducers/progress-button-state';
 import { environment } from '../../environments/environment';
-import { RouterStateUrl } from './router-util';
+import { RouterState } from './router-util';
 import { PrivilagePageSate } from '../models/privilage-page-sate';
 import { isVersionGte } from '../util';
 import {
@@ -44,11 +43,11 @@ import { SelectOptionData } from '../models/select-option-data';
 
 // State interface for the application
 export interface AppState {
+    defaultUserConfiguration: fromDefaultUserConfig.State;
     userConfiguration: fromUserConfig.State;
-    presetKeymaps: Keymap[];
     autoUpdateSettings: autoUpdateSettings.State;
     app: fromApp.State;
-    router: RouterReducerState<RouterStateUrl>;
+    router: RouterReducerState<RouterState>;
     appUpdate: fromAppUpdate.State;
     device: fromDevice.State;
     contributors: fromContributors.State;
@@ -57,8 +56,8 @@ export interface AppState {
 }
 
 export const reducers: ActionReducerMap<AppState> = {
+    defaultUserConfiguration: fromDefaultUserConfig.reducer,
     userConfiguration: fromUserConfig.reducer,
-    presetKeymaps: fromPreset.reducer,
     autoUpdateSettings: autoUpdateSettings.reducer,
     app: fromApp.reducer,
     router: routerReducer,
@@ -406,3 +405,9 @@ export const getFirmwareUpgradeState = createSelector(runningInElectron, getStat
             recoveryModules: []
         };
     });
+
+export const defaultUserConfigState = (state: AppState) => state.defaultUserConfiguration;
+export const getDefaultUserConfigurationKeymaps = createSelector(
+    defaultUserConfigState, fromDefaultUserConfig.getDefaultUserConfigurationKeymaps);
+export const getSelectedAddKeymap = createSelector(
+    defaultUserConfigState, fromDefaultUserConfig.getSelectedKeymap);
